+++ lib/output_workflows/responses/workflow_result.rb
@@ -3,24 +3,34 @@
 module OutputWorkflows
   module Responses
     class WorkflowResult
-      attr_reader :workflow_id, :output, :trace
+      attr_reader :workflow_id, :output, :trace, :aggregations, :attributes
 
-      def initialize(workflow_id:, output: nil, trace: nil)
+      def initialize(workflow_id:, output: nil, trace: nil, aggregations: nil, attributes: nil)
         @workflow_id = workflow_id
         @output = output
         @trace = trace
+        @aggregations = aggregations
+        @attributes = attributes
       end
 
       def self.from_hash(hash)
         new(
-          workflow_id: hash["workflowId"],
-          output: hash["output"],
-          trace: hash["trace"]
+          workflow_id:  hash["workflowId"],
+          output:       hash["output"],
+          trace:        hash["trace"],
+          aggregations: hash["aggregations"],
+          attributes:   hash["attributes"]
         )
       end
 
       def to_h
-        { workflow_id: workflow_id, output: output, trace: trace }
+        {
+          workflow_id:  workflow_id,
+          output:       output,
+          trace:        trace,
+          aggregations: aggregations,
+          attributes:   attributes
+        }
       end
 
       def to_json(*args)